requirements.txt
langchain>=0.1.14
langchain-groq>=0.0.1
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
boto3>=1.28.0
email-validator>=2.0.0
loguru>=0.7.0
groq>=0.4.0
